Intrinsic symbols, context-based determination of variables, units, concepts.
Defining built-in symbols for concepts.
Coined the term 'solitaire' for a lone symbol that we have not yet
resolved as a variable, unit name, or concept name.  Only at run time
do we have context to figure out what a symbol means.
Variable entries in the symbol table need numeric range as well as concept.
Starting to execute statements, including remembering numbered expressions.