Implemented assertf pseudo-function for unit-testing the expression formatter.
Added support for simple string literals in the scanner.
The assertf construct looks like a function, but it is actually a
new type of statement so that strings did not need to be introduced
as a type of expression, which would be very disruptive to the rest of the code.